# Minor performance improvements for RouteValueDictionary